requirements.txt
PyExecJS==1.5.1
requests==2.9.1
beautifulsoup4==4.6.3
# 需配置 node 环境